fix(core): make InstantSearch(Ssr) into regular .js files (#605)
* fix(core): make InstantSearch(Ssr) into regular .js files

The reason I did this is to make sure that rollup turns this into valid components (before the template was compiled into the arguments of `createInstantSearchComponent` instead of into its result.

* chore(test): use a global mock of InstantSearch.js root

* chore: fix file lines


* fix(ssr): correct class name

* chore(test): cleaner snapshot
