+++ src/components/InstantSearch.js
@@ -1,0 +1,3 @@
+import { createInstantSearchComponent } from '../util/createInstantSearchComponent';
+
+export default createInstantSearchComponent({ ssr: false });
