--- src/components/InstantSearch.vue
@@ -1,10 +1,0 @@
-<template>
-  <!-- AisInstantSearch is an empty component that will hold other widgets -->
-  <div :class="suit()"><slot /></div>
-</template>
-
-<script>
-import { createInstantSearchComponent } from '../util/createInstantSearchComponent';
-
-export default createInstantSearchComponent({ ssr: false });
-</script>
