+++ src/widgets.js
@@ -13,10 +13,8 @@ export {
 export { default as AisHighlight } from './components/Highlight.vue';
 export { default as AisHits } from './components/Hits.vue';
 export { default as AisHitsPerPage } from './components/HitsPerPage.vue';
-export { default as AisInstantSearch } from './components/InstantSearch.vue';
-export {
-  default as AisInstantSearchSsr,
-} from './components/InstantSearchSsr.vue';
+export { default as AisInstantSearch } from './components/InstantSearch';
+export { default as AisInstantSearchSsr } from './components/InstantSearchSsr';
 export { default as AisInfiniteHits } from './components/InfiniteHits.vue';
 export { default as AisMenu } from './components/Menu.vue';
 export { default as AisMenuSelect } from './components/MenuSelect.vue';
